@@ -10,23 +10,21 @@ use React\EventLoop\Factory;
 
 class UdpServer extends AbstractServer
 {
-    private const BUFFER_SIZE = 65536;
-
     /**
      * @inheritdoc
      */
     public function start(
-        $minPort,
-        $maxPort,
+        int $minPort,
+        int $maxPort,
         callable $onStart,
         callable $onProcess,
         callable $onEvent,
         ?callable $onTerminate = null,
         ?callable $onError = null
-    ) {
+    ): void {
         $started = false;
         $terminated = false;
-        $port = '';
+        $port = null;
 
         $loop = Factory::create();
         $this->server = new UdpStreamServer($loop);
@@ -34,13 +32,24 @@ class UdpServer extends AbstractServer
         $countDown = 3;
         $loop->addPeriodicTimer(
             1,
-            function () use ($onStart, $onProcess, $onTerminate, &$started, &$terminated, &$loop, &$port, &$countDown) {
+            function () use (
+                $onStart,
+                $onProcess,
+                $onTerminate,
+                &$started,
+                &$terminated,
+                &$loop,
+                &$port,
+                &$countDown
+            ): void {
                 if (!$started) {
                     $onStart($port);
                     $started = true;
                 } else {
+                    // @phpstan-ignore-next-line PHPStan doesn't recognize, that the value can change
                     if ($terminated) {
                         $countDown--;
+                        // @phpstan-ignore-next-line PHPStan doesn't recognize, that the value can change
                         if ($countDown < 0) {
                             if ($onTerminate) {
                                 $onTerminate();
@@ -56,7 +65,7 @@ class UdpServer extends AbstractServer
         );
 
         $chunks = [];
-        $this->server->on('data', function ($data) use ($onEvent, $onError, &$chunks) {
+        $this->server->on('data', function (string $data) use ($onEvent, $onError, &$chunks): void {
             $dataDecoded = $this->processData($data, $chunks);
             if ($dataDecoded) {
                 $this->processEvents([$dataDecoded], $onEvent, $onError);
@@ -72,7 +81,7 @@ class UdpServer extends AbstractServer
      * @param array $chunks Array containing already received chunks.
      * @return string Decoded data.
      */
-    private function processData($data, &$chunks)
+    private function processData(string $data, array &$chunks): string
     {
         switch (ord($data[0])) {
             case 0x78:
@@ -95,7 +104,7 @@ class UdpServer extends AbstractServer
                 $index = ord(substr($data, 10, 1));
                 $total =  ord(substr($data, 11, 1));
                 $chunks[base64_encode($chunkId)][] = substr($data, 12);
-                if ($index == ($total - 1)) {
+                if ($index === ($total - 1)) {
                     $dataJoined = '';
                     foreach ($chunks[base64_encode($chunkId)] as $chunk) {
                         $dataJoined .= $chunk;
@@ -107,8 +116,8 @@ class UdpServer extends AbstractServer
                 }
                 break;
             default:
-                // unknown message
-                throw new InvalidMessageException("Unknown message type.", $data);
+                // message is not encoded
+                $dataDecoded = $data;
         }
         return $dataDecoded;
     }
