@@ -20,22 +20,22 @@ abstract class AbstractGelfTest extends TestCase
             self::assertTrue(!empty($event['host']) || ($event['short_message'] === 'No host'));
             $timestamps[] = $event['timestamp'];
             $hosts[] = empty($event['host']) ? 'empty' : $event['host'];
-            if (!empty($event['file'])) {
-                $file = $event['file'];
+            if (!empty($event['_file'])) {
+                $file = $event['_file'];
                 $exception = $event['_exception'];
-                $line = $event['line'];
+                $line = $event['_line'];
             }
         }
         self::assertEquals([
             0 => [
-                'version' => '1.0',
+                'version' => '1.1',
                 'host' => $hosts[0],
                 'short_message' => 'A debug message.',
                 'level' => 7,
                 'timestamp' => $timestamps[0],
             ],
             1 => [
-                'version' => '1.0',
+                'version' => '1.1',
                 'host' => $hosts[1],
                 'short_message' => 'An alert message',
                 'level' => 1,
@@ -45,30 +45,29 @@ abstract class AbstractGelfTest extends TestCase
                 ],
             ],
             2 => [
-                'version' => '1.0',
+                'version' => '1.1',
                 'host' => $hosts[2],
                 'short_message' => 'Exception example',
                 'level' => 0,
                 'timestamp' => $timestamps[2],
                 'full_message' => "Exception: Test exception (0)\n\n#0 {main}\n",
-                'file' => $file,
-                'line' => $line,
                 '_exception' => $exception,
+                '_line' => $line,
+                '_file' => $file,
             ],
             3 => [
-                'version' => '1.0',
+                'version' => '1.1',
                 'host' => $hosts[3],
                 'short_message' => 'Structured message',
                 'level' => 1,
                 'timestamp' => $timestamps[3],
                 'full_message' => 'There was a foo in bar',
-                'facility' => 'example-facility',
                 '_foo' => 'bar',
                 '_bar' => 'baz',
                 '_barKochba' => 15,
             ],
             4 => [
-                'version' => '1.0',
+                'version' => '1.1',
                 'host' => $hosts[4],
                 'short_message' => 'A warning message.',
                 'level' => 4,
@@ -81,7 +80,7 @@ abstract class AbstractGelfTest extends TestCase
                 ],
             ],
             5 => [
-                'version' => '1.0',
+                'version' => '1.1',
                 'host' => $hosts[5],
                 'short_message' =>
                     file_get_contents(__DIR__ . DIRECTORY_SEPARATOR . 'Clients' . DIRECTORY_SEPARATOR . 'bacon.txt'),
